Mods to uLoadWatch and AppCastingMOOSApp superclass (#86)
Post Release 24.8, uLoadWatch has been modified to publish a "loadcast" request to the apps it is watching. Only these apps will generate _ITER_GAP and _ITER_LEN postings. All other apps will not post this information. This change is on conjunction with the AppCastingMOOSApp superclass, the super class for most apps in the MOOS-IvP release. The motivation of this change is to reduce log file bloat and MOOSDB traffic. Neither reduction was deemed critical, but it was inefficient. These two variable postings, for each app, are essentially only published to be consumed by uLoadWatch. If this app is not even being run in a mission, there is arguably no reason for apps be publishing these variables.